Fix Marshal producing data Unmarshal can't parse
As it currently stands, the marshaling functions print a string
containing the default string representation for time.Time, which is a
reading of the time in the format 2006-01-02 15:04:05.999999999 -0700
MST.
However, this input cannot be accepted not even by the Unmarshaler
itself, because it expects an integer, not Go's default format.

There are, of course, concerns in regard to backwards compatibility.
However, as it currently stands, with the default String()
representation, the result can change based on the used Go version - in
fact, in 1.9 the representation changed to account for the introduction
of the monotonic clock. Because the representation has already changed
in the past and has no guarantee not to change in the future, I suggest
it's best to switch to a known format even if it means breaking
compatibility.